backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "ipython>=9.9.0",
    "langchain>=1.2.6",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "openai>=2.16.0",
    "pillow>=12.1.0",
    "pypdf>=6.6.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "spacy>=3.8.11",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
